requirements.txt
databricks-sql-connector
databricks-sdk
streamlit
streamlit-autorefresh
pandas
plotly
